chatbot/requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.1.0
bitsandbytes==0.45.2
blis==1.2.0
Bottleneck==1.4.2
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
ChatterBot==1.2.0
clean-text==0.6.0
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
datasets==3.3.0
dill==0.3.8
dnspython==2.7.0
emoji==1.7.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.70.0
h5py==3.12.1
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
keras==3.8.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mathparse==0.1.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
msgpack-numpy==0.4.8
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
namex==0.0.8
networkx==3.4.2
nltk==3.9.1
numexpr==2.10.2
numpy==2.0.2
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
peft==0.14.0
pillow==11.1.0
pip-autoremove==0.10.0
preshed==3.0.9
propcache==0.2.1
protobuf==5.29.3
psutil==7.0.0
pyarrow==19.0.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.11.1
pyparsing==3.2.1
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruff==0.9.6
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.38
srsly==2.5.1
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
thinc==8.3.4
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
trl==0.15.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
Unidecode==1.3.8
urllib3==2.3.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
